Fix 64-bit support

+++ RSColorPicker/ColorPickerClasses/RSColorFunctions.h
@@ -12,7 +12,7 @@ BMPixel RSPixelFromHSV(CGFloat H, CGFloat S, CGFloat V);
 void RSHSVFromPixel(BMPixel pixel, CGFloat *h, CGFloat *s, CGFloat *v);
 
 // four floats will be placed into `components`
-void RSGetComponentsForColor(float *components, UIColor *color);
+void RSGetComponentsForColor(CGFloat *components, UIColor *color);
 
 UIImage * RSUIImageWithScale(UIImage *img, CGFloat scale);
